update AuthProvider to use a strategy pattern
